edit

+++ exam_02/ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/27 18:39:32 by hkubo             #+#    #+#             */
-/*   Updated: 2021/05/22 11:26:22 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/05/22 11:26:34 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -577,7 +577,7 @@ int ft_printf_per(const char *arg, int *i, va_list *ap)
 		hex_num = va_arg(*ap, unsigned int);
 		if (!(str_num = ft_itoa_hex(hex_num, "0123456789abcdef")))
 			return (-1);
-		len = print_digit(flag_list, str_num, hex_num, ft_strlen(hex_num));
+		len = print_digit(flag_list, str_num, hex_num, (int)ft_strlen(hex_num));
 	}
 	return (len);
 }
